# PERF: allow creation of read-only array views of scalars