In ufuncs in particular, a bottleneck for dealing with scalars is turning them into arrays.  This PR adds a short-cut for those classes in PyArray_FromAny, where if the array is not specifically requested to be writeable, a readable view of the scalar is returned.  For python scalars, only float is supported; int and complex are more tricky (strings and bytes should presumably be possible, but I didn't add them yet).

Timings:
```                                                                                                                                         
a = np.float64(1.5)                                                                                                                         
%timeit np.array(a, copy=False)  # Comparing with copy=True on main.                                                                        
167->110 ns for float64, 94 ns for float                                                                                                    
%timeit np.add(a, a)                                                                                                                        
512->431 ns for float64,                                                                                                                    
                                                                                                                                            
a = 1.5                                                                                                                                     
%timeit np.array(a, copy=False)  # Comparing with copy=True on main.                                                                        
167->110 ns for float64, 94 ns for float                                                                                                    
%timeit np.add(a, a)                                                                                                                        
526->414 ns                                                                                                                                 
```                                                                                                                                         
An annoyance is that quite a bit of code relies on `np.asanyarray(scalar)` to return a writeable array.  As implemented, this is ensured by checking the `PyArray_FORCECAST` flag, which can be safely unset for `copy=False`. As a result, however, some of the error messages change in nature. This could be resolved with a new flag, if needed. Alternatively, this whole machinery could just be moved to `convert_ufunc_arguments`, since those arguably benefit most from this.