@seberg 2025-10-05
Hmmmm, returning a read-only view seems totally fine to me.  But I am a bit surprised that it makes a 100ns difference.
The allocation for the data should be cached here, so I don't think saving the allocation is actually meaningful, which would mean that `PyArray_Pack` is slow?  But I don't really see why it should be meaningfully slow, since all dtype assignment functions should have a fast-check for their scalar types (I can see that adding 10-20 ns, but 100?).

I am still curious if we can't do fast path for scalar array creation deeper, but I guess I should try that myself a bit.

I am also worried that this very much oversimplifies things to the point of being incorrect.  One thing might be to check that `in_descr` is identical to the actual descriptor.
(One final surprise to me is: this doesn't break `arr[indx]["field"] = val`?! Maybe I am missing it?)