@mhvk 2025-10-05
@seberg - I think the problem is real, but I am not sure this is the best way forward - in particular, I don't quite like the hack I did to make `asanyarray(scalar)` work. I think it would be fair game in `ufunc_object` or perhaps with a flag that explicitly tells that any type of read-only view is OK.

The reason `void` still works is that that check is done before the one I added - if I swap them, it indeed doesn't work (found out the hard way of course!).

Note that I did do timings at various places, and the memory allocation is definitely a problem - though not so much the allocation itself, but rather getting `mem_handler` and tracking the allocation in `PyDataMem_UserNEW`. Specifically, I tried `np.empty(1)`:
- 35 ns for entering array_empty
- ! 64 ns (+30) after parser
- 76 ns before tp_alloc
- ! 104 ns (+30) after tp_alloc (unavoidable?)
- 122 ns before mem_handler
- 139 ns just before data allocation
- !! 201 ns (+60) after data alloc (maybe excess?)
- 202 ns after updateflags
- 221 ns prior to exit

By replacing that part with a simple `malloc` for small sizes, this went down to 138 ns.

With my laptop in performance mode to try to make it more stable (so everything is faster),
```
%timeit np.empty(1)
122->82 ns
%timeit np.array(1.0)
133->92 ns???  (no tracing via change in alloc.c: 117 ns)
%timeit np.add(1.0, 2.0)
526->429 ns
```

I think on second thought this may be the better route. The main annoyance is that one has to deal with different options in `dealloc`, where a warning is raised if no `mem_handler` is available (that warning says one should go through base, which is how I ended up with the current PR...). Maybe another option for small allocations (or at least for scalars) is to use `tp_alloc(type, extra_items)`.