@mhvk 2025-10-05
OK, at least I can reproduce it, it has to do with being on python 3.11. It is also clear it has something to do with subclassing.