@mhvk 2025-10-07
> I was prodding it a bit myself @ngoldbaum once suggest `samply` (which is pretty nice, as it drops you into a browser immediately with just `samply -r 10000 -- spin python perf.py` and you can remove the stuff you don't want.)

Where can I find this `samply`? (searching gives me audio samplers/streamers or some distribution sampler...)