@seberg 2025-10-07
Hah, true it's a bit hidden, conda seems to have it https://github.com/mstange/samply